src/main/java/com/api/MedTrackAPI/service/OrderService.java
package com.api.MedTrackAPI.service;

import com.api.MedTrackAPI.model.Order;
import com.api.MedTrackAPI.repository.OrderRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.ArrayList;
import java.util.List;

@Service
public class OrderService {
    @Autowired
    private final OrderRepository orderRepository;

    public OrderService(OrderRepository orderRepository) {
        this.orderRepository = orderRepository;
    }

    public List<Order>  generateReorderSuggestions(){
        //implement the logic to calc stock levels, compare with thresholds,
        // and generate reorder suggestions
        //return list of orders as reorder suggestions
        // Calculate stock levels based on predefined thresholds
        // Generate reorder suggestions
        List<Order> orders = orderRepository.findAll();

        // Implement the logic to calculate stock levels, compare with thresholds, and generate reorder suggestions
        // Return a list of orders as reorder suggestions

        return orders; // Placeholder, replace with actual logic
    }


}
